# backend/API.yaml
openapi: 3.0.0
info:
  title: JJ Crawler API Documentation
  description: |
    本文档详细介绍晋江文学城爬虫后端系统的所有API接口，已完成T4.4 API实现层的开发，所有接口均已连接到真实的Service层实现。
    
    **API基础信息**
    - **基础URL**: `http://localhost:8000/api/v1`
    - **数据格式**: JSON
    - **认证方式**: 无（开发阶段）
    - **API版本**: v1
    - **实现状态**: ✅ T4.4完成 - 所有Mock API已替换为真实实现
  version: 1.2.0 # 根据您的文档，我使用了最新版本
servers:
  - url: http://localhost:8000/api/v1
    description: JJ Crawler API v1
...
paths:
...
  /crawl/jiazi:
    post:
      summary: 触发夹子榜爬取
      operationId: triggerJiaziCrawl
      tags:
        - Crawler Management
      description: 手动触发夹子榜数据爬取任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                immediate:
                  type: boolean
                  example: true
                metadata:
                  type: object
                  properties:
                    trigger_source:
                      type: string
                      example: manual
                    description:
                      type: string
                      example: 手动触发的夹子榜爬取
      responses:
        "200":
          description: 成功创建爬取任务
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: jiazi_20240101_120000_abc123
                  message:
                    type: string
                    example: 夹子榜爬取任务已创建
                  estimated_duration:
                    type: string
                    example: 2-3分钟
                  status:
                    type: string
                    example: pending
  /crawl/page/{channel}:
    post:
      summary: 触发分类页面爬取
      operationId: triggerChannelCrawl
      tags:
        - Crawler Management
      description: 触发指定分类页面的爬取任务
      parameters:
        - in: path
          name: channel
          schema:
            type: string
          required: true
          description: 频道标识，如 "yq", "ca"
          example: yq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                immediate:
                  type: boolean
                  example: true
                metadata:
                  type: object
                  properties:
                    trigger_source:
                      type: string
                      example: manual
      responses:
        "200":
          description: 成功创建分类页面爬取任务
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: page_yq_20240101_120000_def456
                  message:
                    type: string
                    example: 分类页面爬取任务已创建
                  channel:
                    type: string
                    example: yq
                  estimated_duration:
                    type: string
                    example: 1-2分钟
  /crawl/tasks:
    get:
      summary: 获取任务状态
      operationId: getCrawlTasksStatus
      tags:
        - Crawler Management
      description: 获取所有爬取任务的状态信息
      parameters:
        - in: query
          name: status
          schema:
            type: string
            default: all
            enum:
              - pending
              - running
              - completed
              - failed
              - all
          required: false
          description: 任务状态：pending,running,completed,failed,all
          example: running
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          required: false
          description: 返回数量限制
          example: 20
      responses:
        "200":
          description: 成功获取任务状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        task_id:
                          type: string
                          example: jiazi_20240101_120000_abc123
                        task_type:
                          type: string
                          example: jiazi
                        status:
                          type: string
                          example: running
                        progress:
                          type: integer
                          example: 75
                        items_crawled:
                          type: integer
                          example: 38
                        total_items:
                          type: integer
                          example: 50
                        created_at:
                          type: string
                          format: date-time
                          example: 2024-01-01T12:00:00Z
                        started_at:
                          type: string
                          format: date-time
                          example: 2024-01-01T12:00:30Z
                  recent_completed:
                    type: array
                    items:
                      type: object
                      properties:
                        task_id:
                          type: string
                          example: jiazi_20240101_110000_xyz789
                        task_type:
                          type: string
                          example: jiazi
                        status:
                          type: string
                          example: completed
                        items_crawled:
                          type: integer
                          example: 50
                        completed_at:
                          type: string
                          format: date-time
                          example: 2024-01-01T11:05:00Z
                        duration:
                          type: string
                          example: 4分32秒
                  statistics:
                    type: object
                    properties:
                      total_running:
                        type: integer
                        example: 1
                      total_completed_today:
                        type: integer
                        example: 5
                      total_failed_today:
                        type: integer
                        example: 0
                      success_rate:
                        type: string
                        example: 100%
  /crawl/tasks/{task_id}:
    get:
      summary: 获取单个任务详情
      operationId: getCrawlTaskDetail
      tags:
        - Crawler Management
      description: 获取指定任务的详细信息
      parameters:
        - in: path
          name: task_id
          schema:
            type: string
          required: true
          description: 任务ID
          example: jiazi_20240101_120000_abc123
      responses:
        "200":
          description: 成功获取任务详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: jiazi_20240101_120000_abc123
                  task_type:
                    type: string
                    example: jiazi
                  status:
                    type: string
                    example: completed
                  progress:
                    type: integer
                    example: 100
                  items_crawled:
                    type: integer
                    example: 50
                  created_at:
                    type: string
                    format: date-time
                    example: 2024-01-01T12:00:00Z
                  started_at:
                    type: string
                    format: date-time
                    example: 2024-01-01T12:00:30Z
                  completed_at:
                    type: string
                    format: date-time
                    example: 2024-01-01T12:05:02Z
                  duration:
                    type: string
                    example: 4分32秒
                  metadata:
                    type: object
                    properties:
                      trigger_source:
                        type: string
                        example: scheduled
                      books_new:
                        type: integer
                        example: 3
                      books_updated:
                        type: integer
                        example: 47
                      snapshots_created:
                        type: integer
                        example: 50
                  error_message:
                    type: string
                    nullable: true
                    example: null
...